+++ core/src/main/java/org/bookwoori/core/domain/record/service/RecordService.java
@@ -1,5 +1,7 @@
 package org.bookwoori.core.domain.record.service;
 
+import static org.bouncycastle.asn1.x500.style.RFC4519Style.member;
+
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
@@ -17,6 +19,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class RecordService {
 
     private final RecordRepository recordRepository;
@@ -33,37 +36,46 @@ public class RecordService {
     }
 
 
-    @Transactional
     public Record saveRecord(Record record) {
         return recordRepository.save(record);
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public Record getRecordById(Long recordId) {
         return recordRepository.findById(recordId)
             .orElseThrow(() -> new CustomException(ErrorCode.RECORD_NOT_FOUND));
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public Record getRecordByMemberAndBook(Member member, Book book) {
         return recordRepository.findByMemberAndBook(member, book)
             .orElseThrow(() -> new CustomException(ErrorCode.RECORD_NOT_FOUND));
     }
 
-    @Transactional
     public void deleteRecord(Long recordId) {
+        if (!recordRepository.existsById(recordId)) {
+            throw new CustomException(ErrorCode.RECORD_NOT_FOUND);
+        }
         recordRepository.deleteById(recordId);
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public List<Record> getRecordsByStatus(ReadingStatus status) {
         return recordRepository.findAllByStatus(status);
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public List<Record> getRecordsByMember(Member member) {
         return recordRepository.findAllByMember(member);
     }
 
+    @Transactional(readOnly = true)
+    public boolean existsByMemberAndBook(Member currentMember, Book book) {
+        return recordRepository.existsByMemberAndBook(currentMember, book);
+    }
 
+    @Transactional(readOnly = true)
+    public Optional<Record> getRecordOptByMemberAndBook(Member currentMember, Book book) {
+        return recordRepository.findByMemberAndBook(currentMember, book);
+    }
 }
